Debugging Changes
i)Added fluent Validation
ii)Fixed issues in Book Ticket
iii)Made Some DB Schema Change

+++ TrainTravel.API/Controllers/BookingController.cs
@@ -16,12 +16,17 @@ namespace TrainTravel.API.Controllers
         private readonly ILogger<BookingController> logger;
         private readonly IMapper mapper;
         private readonly IBookRepository bookRepository;
+        private readonly ITicketRepository ticketRepository;
+        private readonly IPassengerRepository passengerRepository;
 
-        public BookingController(ILogger<BookingController> logger, IMapper mapper,IBookRepository bookRepository)
+        public BookingController(ILogger<BookingController> logger, IMapper mapper,IBookRepository bookRepository,
+            ITicketRepository ticketRepository,IPassengerRepository passengerRepository)
         {
             this.logger = logger;
             this.mapper = mapper;
             this.bookRepository = bookRepository;
+            this.ticketRepository = ticketRepository;
+            this.passengerRepository = passengerRepository;
         }
         //POST :/api/booking/book
         [HttpPost]
@@ -31,13 +36,30 @@ namespace TrainTravel.API.Controllers
         {
             try
             {             
+                List<PassengerResponseDto> response =new List<PassengerResponseDto>();
                 var user = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
                 var bookingsDomainModel = mapper.Map<BookingsData>(bookingRequestDto);
                 bookingsDomainModel.UserId = user;
                 var bookingResult = await bookRepository.BookAsync(bookingsDomainModel);
                 if(bookingResult!=null && bookingResult.Id != Guid.Empty) 
                 {
-                    
+                    var ticketDomainModel = mapper.Map<TicketData>(bookingRequestDto);
+                    ticketDomainModel.BookingId= bookingResult.Id;
+                    var ticketResult = await ticketRepository.CreateTicketAsync(ticketDomainModel);           
+                    if (ticketResult != null && ticketResult.Id != Guid.Empty)
+                    {
+                        foreach (var passenger in bookingRequestDto.passengerRequests)
+                        {
+                            var passengerDomainModel = mapper.Map<PassengerData>(passenger);
+                            passengerDomainModel.TicketId=ticketDomainModel.Id;
+                            var passengerResult=await passengerRepository.CreatePassengerAsync(passengerDomainModel);
+                            if(passengerResult!=null)
+                            {
+
+                                response.Add(mapper.Map<PassengerResponseDto>(passengerResult));
+                            }
+                        }
+                    }
                 }
                 return Ok("");
             }
@@ -57,7 +79,7 @@ namespace TrainTravel.API.Controllers
             try
             {           
                 id=id.ToLower();
-                var result = await bookRepository.GetAllBookingForUserAsync(id);               
+                var result = await bookRepository.GetAllBookingForUserAsync(id);              
                 return Ok(mapper.Map<List<BookingRequestDto>>(result));
             }
             catch (Exception ex)
@@ -72,16 +94,28 @@ namespace TrainTravel.API.Controllers
         [HttpDelete]
         [Route("{id:Guid}")]
         [Authorize(Roles = "User,Admin")]
-        public async Task<IActionResult> Delete([FromRoute] Guid id)
+        public async Task<IActionResult> CancelTicket([FromRoute] Guid id)
         {
-            var bookingDomain = await bookRepository.DeleteBookingAsync(id);
-            if (bookingDomain == null)
+            var user = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var passengerRemoved = await passengerRepository.DeletePassengerAsync(id);
+            if (passengerRemoved != null && passengerRemoved.TicketId != Guid.Empty)
             {
-                return NotFound();
+                var passengerRemovedTicketId = passengerRemoved.TicketId;
+                var passengerForTicket = await passengerRepository.GetAllPassengersTicketAsync(passengerRemovedTicketId);
+                if (passengerForTicket != null && passengerForTicket.Count == 0)
+                {
+                    var ticketRemoved = await ticketRepository.DeleteTicketAsync(passengerRemovedTicketId);
+                    if(ticketRemoved!=null && ticketRemoved.Id!=Guid.Empty)
+                    {
+                        var bookingDomain = await bookRepository.DeleteBookingAsync(id);
+                        if (bookingDomain != null)
+                        {
+                            return Ok("Successfully Cancelled the Ticket");
+                        }
+                    }
+                }                
             }
-
-            return Ok(mapper.Map<BookingRequestDto>(bookingDomain));
-
+            return NotFound();     
         }
     }
 }
